test(item-sliding): migrate async
